+++ rustsbi/src/ecall/rfence.rs
@@ -1,4 +1,7 @@
 use super::SbiRet;
+use crate::rfence;
+use crate::hart_mask::HartMask;
+use crate::ipi::max_hart_id;
 
 const FUNCTION_RFENCE_REMOTE_FENCE_I: usize = 0x0;
 const FUNCTION_RFENCE_REMOTE_SFENCE_VMA: usize = 0x1;
@@ -22,51 +25,84 @@ pub fn handle_ecall_rfence(function: usize, param0: usize, param1: usize, param2
     }
 }
 
+// If None = max_hart_id(), that means IPI extension is not supported.
+// In RustSBI, RFENCE support requires an IPI support is implemented.
+// If platform does not provide IPI support, RustSBI will disable RFENCE
+// interface access from supervisor level.
+
 #[inline]
 fn remote_fence_i(hart_mask: usize, hart_mask_base: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_fence_i(hart_mask)
 }
 
 #[inline]
 fn remote_sfence_vma(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_sfence_vma(hart_mask, start_addr, size)
 }
 
 #[inline]
 fn remote_sfence_vma_asid(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize, asid: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size, asid));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_sfence_vma_asid(hart_mask, start_addr, size, asid)
 }
 
 #[inline]
 fn remote_hfence_gvma_vmid(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize, vmid: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size, vmid));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_hfence_gvma_vmid(hart_mask, start_addr, size, vmid)
 }
 
 #[inline]
 fn remote_hfence_gvma(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_hfence_gvma(hart_mask, start_addr, size)
 }
 
 #[inline]
 fn remote_hfence_vvma_asid(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize, asid: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size, asid));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_hfence_vvma_asid(hart_mask, start_addr, size, asid)
 }
 
 #[inline]
 fn remote_hfence_vvma(hart_mask: usize, hart_mask_base: usize, start_addr: usize, size: usize) -> SbiRet {
-    // todo
-    drop((hart_mask, hart_mask_base, start_addr, size));
-    SbiRet::not_supported()
+    let max_hart_id = if let Some(id) = max_hart_id() {
+        id
+    } else {
+        return SbiRet::not_supported()
+    };
+    let hart_mask = unsafe { HartMask::from_addr(hart_mask, hart_mask_base, max_hart_id) };
+    rfence::remote_hfence_vvma(hart_mask, start_addr, size)
 }